+++ src/minigame/brawls/time/TimeBrawlFactory.sol
@@ -25,21 +25,18 @@ contract TimeBrawlFactory is ITimeBrawlFactory, Initializable {
     error BrawlAlreadyExists(uint256 id, address brawl);
     error SameBaseQuote(address base, address quote);
 
-    constructor() {
-    }
+    constructor() {}
 
-    function createBrawl(
-        address base,
-        address quote,
-        uint256 startPrice_,
-        address bet_,
-        uint256 endTime_
-    ) external override returns (address orderbook) {
+    function createBrawl(address base, address quote, uint256 startPrice_, address bet_, uint256 endTime_)
+        external
+        override
+        returns (address orderbook)
+    {
         if (msg.sender != portal) {
             revert InvalidAccess(msg.sender, portal);
         }
 
-        uint id = allBrawlsLength() + 1;
+        uint256 id = allBrawlsLength() + 1;
 
         address brawl = _predictAddress(base, quote, id);
 
@@ -54,17 +51,8 @@ contract TimeBrawlFactory is ITimeBrawlFactory, Initializable {
             revert BrawlAlreadyExists(id, brawl);
         }
 
-        address proxy = CloneFactory._createCloneWithSalt(
-            impl,
-            _getSalt(base, quote, id)
-        );
-        ITimeBrawl(proxy).initialize(
-            id,
-            portal,
-            startPrice_,
-            bet_,
-            endTime_
-        );
+        address proxy = CloneFactory._createCloneWithSalt(impl, _getSalt(base, quote, id));
+        ITimeBrawl(proxy).initialize(id, portal, startPrice_, bet_, endTime_);
         allBrawls.push(proxy);
         return (proxy);
     }
@@ -77,10 +65,14 @@ contract TimeBrawlFactory is ITimeBrawlFactory, Initializable {
         return _predictAddress(base, quote, id);
     }
 
-    function getBrawlInfo(address base, address quote, uint256 id) external view override returns (ITimeBrawl.Brawl memory brawl) {
+    function getBrawlInfo(address base, address quote, uint256 id)
+        external
+        view
+        override
+        returns (ITimeBrawl.Brawl memory brawl)
+    {
         return ITimeBrawl(_predictAddress(base, quote, id)).getBrawl();
     }
-
 
     /**
      * @dev Initialize orderbook factory contract with engine address, reinitialize if engine is reset.
@@ -102,27 +94,17 @@ contract TimeBrawlFactory is ITimeBrawlFactory, Initializable {
         bytes32 salt = keccak256(abi.encodePacked("timebrawl", version));
         assembly {
             addr := create2(0, add(bytecode, 0x20), mload(bytecode), salt)
-            if iszero(extcodesize(addr)) {
-                revert(0, 0)
-            }
+            if iszero(extcodesize(addr)) { revert(0, 0) }
         }
         impl = addr;
     }
 
-    function _predictAddress(
-        address base,
-        address quote,
-        uint256 id
-    ) internal view returns (address) {
+    function _predictAddress(address base, address quote, uint256 id) internal view returns (address) {
         bytes32 salt = _getSalt(base, quote, id);
         return CloneFactory.predictAddressWithSalt(address(this), impl, salt);
     }
 
-    function _getSalt(
-        address base,
-        address quote,
-        uint256 id
-    ) internal pure returns (bytes32) {
+    function _getSalt(address base, address quote, uint256 id) internal pure returns (bytes32) {
         return keccak256(abi.encodePacked(base, quote, id));
     }
 }
