+++ src/safu/Coupon.sol
@@ -7,8 +7,8 @@ import "@openzeppelin/contracts/access/AccessControl.sol";
 import "./interfaces/ICoupon.sol";
 import "./interfaces/INetworkState.sol";
 
-contract Coupon is ERC721, AccessControl  {
-     // Create a new role identifier for the minter role
+contract Coupon is ERC721, AccessControl {
+    // Create a new role identifier for the minter role
     bytes32 public constant MINTER_ROLE = keccak256("MINTER_ROLE");
     bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
 
@@ -18,8 +18,7 @@ contract Coupon is ERC721, AccessControl  {
         return super.supportsInterface(interfaceId);
     }
 
-    constructor(address manager_)
-    ERC721("SAFU BOND COUPON", "COUPON") {
+    constructor(address manager_) ERC721("SAFU BOND COUPON", "COUPON") {
         _grantRole(DEFAULT_ADMIN_ROLE, _msgSender());
 
         _grantRole(MINTER_ROLE, _msgSender());
@@ -27,13 +26,13 @@ contract Coupon is ERC721, AccessControl  {
         _grantRole(BURNER_ROLE, _msgSender());
         manager = manager_;
     }
-    
+
     function setManager(address manager_) public {
         require(hasRole(DEFAULT_ADMIN_ROLE, msg.sender), "MTRV1: Caller is not a default admin");
         manager = manager_;
     }
 
-    function mint(address to, uint256 tokenId_) external  {
+    function mint(address to, uint256 tokenId_) external {
         // Check that the calling account has the minter role
         require(hasRole(MINTER_ROLE, msg.sender), "MTRV1: Caller is not a minter");
         _mint(to, tokenId_);
@@ -45,9 +44,8 @@ contract Coupon is ERC721, AccessControl  {
         _burn(tokenId_);
     }
 
-    function burnFromVault(uint vaultId_) external{
-        require(INetworkState(manager).getVault(vaultId_)  == _msgSender(), "MTRV1: Caller is not vault");
+    function burnFromVault(uint256 vaultId_) external {
+        require(INetworkState(manager).getVault(vaultId_) == _msgSender(), "MTRV1: Caller is not vault");
         _burn(vaultId_);
     }
 }
-
