+++ src/safu/interfaces/IERC20Minimal.sol
@@ -3,7 +3,7 @@
 pragma solidity >=0.5.0;
 
 interface IERC20Minimal {
-    function totalSupply() external view returns (uint);
-    function balanceOf(address owner) external view returns (uint);
+    function totalSupply() external view returns (uint256);
+    function balanceOf(address owner) external view returns (uint256);
     function mint(address to, uint256 amount) external;
 }
